Serve entities properly again
